+++ src/kernel_evm/kernel/src/block_in_progress.rs
@@ -171,13 +171,7 @@ impl BlockInProgress {
         gas_price: U256,
         transactions: VecDeque<Transaction>,
     ) -> BlockInProgress {
-        Self::new_with_ticks(
-            number,
-            H256::zero(),
-            gas_price,
-            transactions,
-            tick_model::top_level_overhead_ticks(),
-        )
+        Self::new_with_ticks(number, H256::zero(), gas_price, transactions, 0u64)
     }
 
     pub fn from_queue_element(