+++ src/kernel_evm/CHANGES.md
@@ -15,6 +15,7 @@
 - Simulate if a transaction has a correct chain id. (!9752)
 - Fees are now paid by the sender. (!9480)
 - Unused gas is refunded after succesful transactions. (!9915)
+- Impose a maximum gas limit per transaction based on the tick model. (!10094)
 
 ### EVM Node